fix(attribute): Better naming prefix `CanBe` for boolean attributes.
