+++ src/Form/CanBeReadonly.php
@@ -15,7 +15,7 @@ use UIAwesome\Html\Attribute\Values\Attribute;
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
-trait HasReadonly
+trait CanBeReadonly
 {
     /**
      * Sets the `readonly` attribute.
@@ -23,7 +23,6 @@ trait HasReadonly
      * Usage example:
      * ```php
      * $element->readonly(true);
-     * $element->readonly(false);
      * $element->readonly(null);
      * ```
      *
