+++ tests/CanBeDisabledTest.php
@@ -2,37 +2,37 @@
 
 declare(strict_types=1);
 
-namespace UIAwesome\Html\Attribute\Tests\Form;
+namespace UIAwesome\Html\Attribute\Tests;
 
 use PHPUnit\Framework\Attributes\{DataProviderExternal, Group};
 use PHPUnit\Framework\TestCase;
-use UIAwesome\Html\Attribute\Form\HasChecked;
-use UIAwesome\Html\Attribute\Tests\Provider\Form\CheckedProvider;
+use UIAwesome\Html\Attribute\CanBeDisabled;
+use UIAwesome\Html\Attribute\Tests\Provider\DisabledProvider;
 use UIAwesome\Html\Attribute\Values\Attribute;
 use UIAwesome\Html\Helper\Attributes;
 use UIAwesome\Html\Mixin\HasAttributes;
 
 /**
- * Unit tests for the {@see HasChecked} trait managing the `checked` HTML attribute.
+ * Unit tests for the {@see CanBeDisabled} trait managing the `disabled` HTML attribute.
  *
  * Test coverage.
  * - Ensures fluent setters return new instances (immutability).
- * - Ensures no attributes are set when the `checked` attribute is not provided.
- * - Sets the `checked` HTML attribute and renders the expected output.
+ * - Ensures no attributes are set when the `disabled` attribute is not provided.
+ * - Sets the `disabled` HTML attribute and renders the expected output.
  *
- * {@see CheckedProvider} for test case data providers.
+ * {@see DisabledProvider} for test case data providers.
  *
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 #[Group('attribute')]
-final class HasCheckedTest extends TestCase
+final class CanBeDisabledTest extends TestCase
 {
-    public function testReturnEmptyWhenCheckedAttributeNotSet(): void
+    public function testReturnEmptyWhenDisabledAttributeNotSet(): void
     {
         $instance = new class {
+            use CanBeDisabled;
             use HasAttributes;
-            use HasChecked;
         };
 
         self::assertEmpty(
@@ -41,16 +41,16 @@ final class HasCheckedTest extends TestCase
         );
     }
 
-    public function testReturnNewInstanceWhenSettingCheckedAttribute(): void
+    public function testReturnNewInstanceWhenSettingDisabledAttribute(): void
     {
         $instance = new class {
+            use CanBeDisabled;
             use HasAttributes;
-            use HasChecked;
         };
 
         self::assertNotSame(
             $instance,
-            $instance->checked(null),
+            $instance->disabled(null),
             'Should return a new instance when setting the attribute, ensuring immutability.',
         );
     }
@@ -58,24 +58,24 @@ final class HasCheckedTest extends TestCase
     /**
      * @phpstan-param mixed[] $attributes
      */
-    #[DataProviderExternal(CheckedProvider::class, 'values')]
-    public function testSetCheckedAttributeValue(
-        bool|null $checked,
+    #[DataProviderExternal(DisabledProvider::class, 'values')]
+    public function testSetDisabledAttributeValue(
+        bool|null $disabled,
         array $attributes,
-        bool|string|null $expectedValue,
+        bool|null $expectedValue,
         string $expectedRenderAttributes,
         string $message,
     ): void {
         $instance = new class {
+            use CanBeDisabled;
             use HasAttributes;
-            use HasChecked;
         };
 
-        $instance = $instance->attributes($attributes)->checked($checked);
+        $instance = $instance->attributes($attributes)->disabled($disabled);
 
         self::assertSame(
             $expectedValue,
-            $instance->getAttribute(Attribute::CHECKED, ''),
+            $instance->getAttribute(Attribute::DISABLED, ''),
             $message,
         );
         self::assertSame(
