+++ tests/Form/CanBeReadonlyTest.php
@@ -6,33 +6,33 @@ namespace UIAwesome\Html\Attribute\Tests\Form;
 
 use PHPUnit\Framework\Attributes\{DataProviderExternal, Group};
 use PHPUnit\Framework\TestCase;
-use UIAwesome\Html\Attribute\Form\HasMultiple;
-use UIAwesome\Html\Attribute\Tests\Provider\Form\MultipleProvider;
+use UIAwesome\Html\Attribute\Form\CanBeReadonly;
+use UIAwesome\Html\Attribute\Tests\Provider\Form\ReadonlyProvider;
 use UIAwesome\Html\Attribute\Values\Attribute;
 use UIAwesome\Html\Helper\Attributes;
 use UIAwesome\Html\Mixin\HasAttributes;
 
 /**
- * Unit tests for the {@see HasMultiple} trait managing the `multiple` HTML attribute.
+ * Unit tests for the {@see CanBeReadonly} trait managing the `readonly` HTML attribute.
  *
  * Test coverage.
  * - Ensures fluent setters return new instances (immutability).
- * - Ensures no attributes are set when the `multiple` attribute is not provided.
- * - Sets the `multiple` HTML attribute and renders the expected output.
+ * - Ensures no attributes are set when the `readonly` attribute is not provided.
+ * - Sets the `readonly` HTML attribute and renders the expected output.
  *
- * {@see MultipleProvider} for test case data providers.
+ * {@see ReadonlyProvider} for test case data providers.
  *
  * @copyright Copyright (C) 2026 Terabytesoftw.
  * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
  */
 #[Group('attribute')]
-final class HasMultipleTest extends TestCase
+final class CanBeReadonlyTest extends TestCase
 {
-    public function testReturnEmptyWhenMultipleAttributeNotSet(): void
+    public function testReturnEmptyWhenReadonlyAttributeNotSet(): void
     {
         $instance = new class {
+            use CanBeReadonly;
             use HasAttributes;
-            use HasMultiple;
         };
 
         self::assertEmpty(
@@ -41,16 +41,16 @@ final class HasMultipleTest extends TestCase
         );
     }
 
-    public function testReturnNewInstanceWhenSettingMultipleAttribute(): void
+    public function testReturnNewInstanceWhenSettingReadonlyAttribute(): void
     {
         $instance = new class {
+            use CanBeReadonly;
             use HasAttributes;
-            use HasMultiple;
         };
 
         self::assertNotSame(
             $instance,
-            $instance->multiple(null),
+            $instance->readonly(null),
             'Should return a new instance when setting the attribute, ensuring immutability.',
         );
     }
@@ -58,24 +58,24 @@ final class HasMultipleTest extends TestCase
     /**
      * @phpstan-param mixed[] $attributes
      */
-    #[DataProviderExternal(MultipleProvider::class, 'values')]
-    public function testSetMultipleAttributeValue(
-        bool|null $multiple,
+    #[DataProviderExternal(ReadonlyProvider::class, 'values')]
+    public function testSetReadonlyAttributeValue(
+        bool|null $readonly,
         array $attributes,
-        bool|string|null $expectedValue,
+        bool|null $expectedValue,
         string $expectedRenderAttributes,
         string $message,
     ): void {
         $instance = new class {
+            use CanBeReadonly;
             use HasAttributes;
-            use HasMultiple;
         };
 
-        $instance = $instance->attributes($attributes)->multiple($multiple);
+        $instance = $instance->attributes($attributes)->readonly($readonly);
 
         self::assertSame(
             $expectedValue,
-            $instance->getAttribute(Attribute::MULTIPLE, ''),
+            $instance->getAttribute(Attribute::READONLY, ''),
             $message,
         );
         self::assertSame(
